requirements.txt
PyQt5
pyqt5-plugins
PyQt5-Qt5
PyQt5-sip
PyQt5-stubs
pyqt5-tools
pyqtgraph
pygame
pydub
mutagen
numpy
pypresence
aiodns
fire
pycares
qasync
asyncio
pyimgur
requests
python-dotenv
psutil
packaging
Pillow